{{cookiecutter.root_folder}}/requirements.txt
cryptography
azure-batch
azure-storage
azure-datalake-store
azure-mgmt-resource
